PixelMinds/nlp-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.0
spacy==3.7.5
aiofiles==24.1.0
python-dotenv==1.0.1
sentence-transformers==3.0.0
